made some updation like admin will not be able to delete the features and facilities which are allocated to rooms

+++ admin/ajax/features_facilities.php
@@ -43,16 +43,25 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        if($res)
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],"i");
+        if(mysqli_num_rows($check_q) == 0)
         {
-            echo 1;
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values,'i');
+            if($res)
+            {
+                echo 1;
+            }
+            else
+            {
+                echo 0;
+            }
         }
         else
         {
-            echo 0;
+            echo 'room_added';
         }
+
     }
 
     if(isset($_POST['add_facility']))
@@ -111,30 +120,40 @@
     {
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facility']];
+        
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facility']],"i");
 
-        // Fetch the icon filename
-        $pre_q = "SELECT `icon` FROM `facilities` WHERE `id`=?";
-        $res = select($pre_q, $values, 'i');
-        $img = mysqli_fetch_assoc($res);
-
-        // Delete the icon file from the server
-        if($img && !empty($img['icon'])) {
-            $icon_path = FACILITIES_FOLDER . $img['icon'];
-            if(file_exists($icon_path)){
-                unlink($icon_path);
-            }
-        }
-
-        // Delete the row from the database
-        $q = "DELETE FROM `facilities` WHERE `id`=?";
-        $res = delete($q, $values, 'i');
-        if($res)
+        if(mysqli_num_rows($check_q)==0)
         {
-            echo 1;
+            // Fetch the icon filename
+            $pre_q = "SELECT `icon` FROM `facilities` WHERE `id`=?";
+            $res = select($pre_q, $values, 'i');
+            $img = mysqli_fetch_assoc($res);
+    
+            // Delete the icon file from the server
+            if($img && !empty($img['icon'])) {
+                $icon_path = FACILITIES_FOLDER . $img['icon'];
+                if(file_exists($icon_path)){
+                    unlink($icon_path);
+                }
+            }
+    
+            // Delete the row from the database
+            $q = "DELETE FROM `facilities` WHERE `id`=?";
+            $res = delete($q, $values, 'i');
+            if($res)
+            {
+                echo 1;
+            }
+            else
+            {
+                echo 0;
+            }   
         }
         else
         {
-            echo 0;
+            echo 'room_added';
         }
+        
     }
 ?>
